fix some builtin problems but need to fix pipe and others

+++ includes/structs.h
@@ -72,7 +72,7 @@ typedef struct s_remove_args
 	char	**new_args;
 	int		i;
 	int		origin_i;
-	int		buffersize;
+	size_t		buffersize;
 }	t_rm_args;
 
 typedef struct s_Redirection
