+++ includes/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: suminkwon <suminkwon@student.42.fr>        +#+  +:+       +#+        */
+/*   By: hlee-sun <hlee-sun@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 17:31:24 by sukwon            #+#    #+#             */
-/*   Updated: 2024/10/10 22:41:33 by suminkwon        ###   ########.fr       */
+/*   Updated: 2024/10/13 17:01:50 by hlee-sun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 # include "minishell.h"
 
 int		set_ehcoctl(int flag);
-char    *ft_realloc_single(char *token, int position, int *bufsize);
-char	**ft_realloc_double(char **token, int position, int *bufsize);
+char    *ft_realloc_single(char *token, size_t position, size_t *bufsize);
+char	**ft_realloc_double(char **token, size_t position, size_t *bufsize);
 char	*ft_strndup(char *str, int n);
 int		ft_isspace(char c);
 int		ft_strcmp(const char *line, const char *limiter);
